+++ python/metatomic_torch/metatomic/torch/ase_calculator.py
@@ -7,7 +7,7 @@ from typing import Dict, List, Optional, Tuple, Union
 import metatensor.torch as mts
 import numpy as np
 import torch
-import vesin
+import vesin.metatomic
 from metatensor.torch import Labels, TensorBlock, TensorMap
 from torch.profiler import record_function
 
@@ -16,22 +16,29 @@ from . import (
     ModelEvaluationOptions,
     ModelMetadata,
     ModelOutput,
+    NeighborListOptions,
     System,
     load_atomistic_model,
     pick_device,
     pick_output,
-    register_autograd_neighbors,
 )
 
 
 import ase  # isort: skip
-import ase.neighborlist  # isort: skip
 import ase.calculators.calculator  # isort: skip
 from ase.calculators.calculator import (  # isort: skip
     InputError,
     PropertyNotImplementedError,
     all_properties as ALL_ASE_PROPERTIES,
 )
+
+
+try:
+    from nvalchemiops.neighborlist import neighbor_list as nvalchemi_neighbor_list
+
+    HAS_NVALCHEMIOPS = True
+except ImportError:
+    HAS_NVALCHEMIOPS = False
 
 FilePath = Union[str, bytes, pathlib.PurePath]
 
@@ -137,9 +144,9 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
 
     Neighbor lists are computed using the fast
     `vesin <https://luthaf.fr/vesin/latest/index.html>`_ neighbor list library,
-    unless the system has mixed periodic and non-periodic boundary conditions (which
-    are not yet supported by ``vesin``), in which case the slower ASE neighbor list
-    is used.
+    either on CPU or GPU depending on the device of the model. If
+    `nvalchemiops <https://github.com/NVIDIA/nvalchemi-toolkit-ops>`_
+    is installed, full neighbor lists on GPU will be computed with it instead.
     """
 
     def __init__(
@@ -419,17 +426,6 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
                 atoms=atoms, dtype=self._dtype, device=self._device
             )
             system = System(types, positions, cell, pbc)
-            # Compute the neighbors lists requested by the model
-            for options in self._model.requested_neighbor_lists():
-                neighbors = _compute_ase_neighbors(
-                    atoms, options, dtype=self._dtype, device=self._device
-                )
-                register_autograd_neighbors(
-                    system,
-                    neighbors,
-                    check_consistency=self.parameters["check_consistency"],
-                )
-                system.add_neighbor_list(options, neighbors)
             # Get the additional inputs requested by the model
             for name, option in self._model.requested_inputs().items():
                 input_tensormap = _get_ase_input(
@@ -437,6 +433,13 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
                 )
                 system.add_data(name, input_tensormap)
             systems.append(system)
+
+        # Compute the neighbors lists requested by the model
+        input_systems = _compute_requested_neighbors(
+            systems=systems,
+            requested_options=self._model.requested_neighbor_lists(),
+            check_consistency=self.parameters["check_consistency"],
+        )
 
         available_outputs = self._model.capabilities().outputs
         for key in outputs:
@@ -449,7 +452,7 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
             selected_atoms=selected_atoms,
         )
         return self._model(
-            systems=systems,
+            systems=input_systems,
             options=options,
             check_consistency=self.parameters["check_consistency"],
         )
@@ -591,27 +594,22 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
         with record_function("MetatomicCalculator::compute_neighbors"):
             # convert from ase.Atoms to metatomic.torch.System
             system = System(types, positions, cell, pbc)
-            for options in self._model.requested_neighbor_lists():
-                neighbors = _compute_ase_neighbors(
-                    atoms, options, dtype=self._dtype, device=self._device
-                )
-                register_autograd_neighbors(
-                    system,
-                    neighbors,
-                    check_consistency=self.parameters["check_consistency"],
-                )
-                system.add_neighbor_list(options, neighbors)
+            input_system = _compute_requested_neighbors(
+                systems=[system],
+                requested_options=self._model.requested_neighbor_lists(),
+                check_consistency=self.parameters["check_consistency"],
+            )[0]
 
         with record_function("MetatomicCalculator::get_model_inputs"):
             for name, option in self._model.requested_inputs().items():
                 input_tensormap = _get_ase_input(
                     atoms, name, option, dtype=self._dtype, device=self._device
                 )
-                system.add_data(name, input_tensormap)
+                input_system.add_data(name, input_tensormap)
 
         # no `record_function` here, this will be handled by AtomisticModel
         outputs = self._model(
-            [system],
+            [input_system],
             run_options,
             check_consistency=self.parameters["check_consistency"],
         )
@@ -667,18 +665,14 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
 
             if calculate_energies:
                 energies_values = energies.block().values.detach().reshape(-1)
-                energies_values = energies_values.to(device="cpu").to(
-                    dtype=torch.float64
-                )
                 atom_indexes = energies.block().samples.column("atom")
-
                 result = torch.zeros_like(energies_values)
                 result.index_add_(0, atom_indexes, energies_values)
-                self.results["energies"] = result.numpy()
+                self.results["energies"] = result.cpu().double().numpy()
 
             if calculate_energy:
                 energy_values = energy.block().values.detach()
-                energy_values = energy_values.to(device="cpu").to(dtype=torch.float64)
+                energy_values = energy_values.cpu().double()
                 self.results["energy"] = energy_values.numpy()[0, 0]
 
             if calculate_forces:
@@ -691,7 +685,7 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
                 else:
                     forces_values = -system.positions.grad
                 forces_values = forces_values.reshape(-1, 3)
-                forces_values = forces_values.to(device="cpu").to(dtype=torch.float64)
+                forces_values = forces_values.cpu().double()
                 self.results["forces"] = forces_values.numpy()
 
             if calculate_stress:
@@ -700,7 +694,7 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
                 else:
                     stress_values = strain.grad / atoms.cell.volume
                 stress_values = stress_values.reshape(3, 3)
-                stress_values = stress_values.to(device="cpu").to(dtype=torch.float64)
+                stress_values = stress_values.cpu().double()
                 self.results["stress"] = _full_3x3_to_voigt_6_stress(
                     stress_values.numpy()
                 )
@@ -780,26 +774,18 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
                 cell = cell @ strain
                 strains.append(strain)
             system = System(types, positions, cell, pbc)
-            # Compute the neighbors lists requested by the model
-            for options in self._model.requested_neighbor_lists():
-                neighbors = _compute_ase_neighbors(
-                    atoms, options, dtype=self._dtype, device=self._device
-                )
-                register_autograd_neighbors(
-                    system,
-                    neighbors,
-                    check_consistency=self.parameters["check_consistency"],
-                )
-                system.add_neighbor_list(options, neighbors)
             systems.append(system)
 
-        options = ModelEvaluationOptions(
-            length_unit="angstrom",
-            outputs=outputs,
-        )
-        predictions = self._model(
+        # Compute the neighbors lists requested by the model
+        input_systems = _compute_requested_neighbors(
             systems=systems,
-            options=options,
+            requested_options=self._model.requested_neighbor_lists(),
+            check_consistency=self.parameters["check_consistency"],
+        )
+
+        predictions = self._model(
+            systems=input_systems,
+            options=ModelEvaluationOptions(length_unit="angstrom", outputs=outputs),
             check_consistency=self.parameters["check_consistency"],
         )
         energies = predictions[self._energy_key]
@@ -807,12 +793,7 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
         if energy_per_atom:
             # Get per-atom energies
             sorted_block = mts.sort_block(energies.block(), axes="samples")
-            energies_values = (
-                sorted_block.values.detach()
-                .reshape(-1)
-                .to(device="cpu")
-                .to(dtype=torch.float64)
-            )
+            energies_values = sorted_block.values.detach().reshape(-1)
 
             split_sizes = [len(system) for system in systems]
             atom_indices = sorted_block.samples.column("atom")
@@ -822,7 +803,9 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
             for atom_indices, values in zip(
                 split_atom_indices, energies_values, strict=True
             ):
-                split_energy = torch.zeros(len(atom_indices), dtype=values.dtype)
+                split_energy = torch.zeros(
+                    len(atom_indices), dtype=values.dtype, device=values.device
+                )
                 split_energy.index_add_(0, atom_indices, values)
                 split_energies.append(split_energy)
 
@@ -831,17 +814,23 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
                 .block()
                 .values.detach()
                 .cpu()
+                .double()
                 .numpy()
                 .flatten()
                 .tolist()
             )
             results_as_numpy_arrays = {
                 "energy": total_energy,
-                "energies": [e.numpy() for e in split_energies],
+                "energies": [e.cpu().double().numpy() for e in split_energies],
             }
         else:
             results_as_numpy_arrays = {
-                "energy": energies.block().values.squeeze(-1).detach().cpu().numpy(),
+                "energy": energies.block()
+                .values.squeeze(-1)
+                .detach()
+                .cpu()
+                .double()
+                .numpy(),
             }
 
         if compute_forces_and_stresses:
@@ -955,79 +944,6 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
         return metatensor_outputs
 
 
-def _compute_ase_neighbors(atoms, options, dtype, device):
-    # options.strict is ignored by this function, since `ase.neighborlist.neighbor_list`
-    # only computes strict NL, and these are valid even with `strict=False`
-
-    if np.all(atoms.pbc) or np.all(~atoms.pbc):
-        nl_i, nl_j, nl_S, nl_D = vesin.ase_neighbor_list(
-            "ijSD",
-            atoms,
-            cutoff=options.engine_cutoff(engine_length_unit="angstrom"),
-        )
-    else:
-        nl_i, nl_j, nl_S, nl_D = ase.neighborlist.neighbor_list(
-            "ijSD",
-            atoms,
-            cutoff=options.engine_cutoff(engine_length_unit="angstrom"),
-        )
-
-    if not options.full_list:
-        # The pair selection code here below avoids a relatively slow loop over
-        # all pairs to improve performance
-        reject_condition = (
-            # we want a half neighbor list, so drop all duplicated neighbors
-            (nl_j < nl_i)
-            | (
-                (nl_i == nl_j)
-                & (
-                    # only create pairs with the same atom twice if the pair spans more
-                    # than one unit cell
-                    ((nl_S[:, 0] == 0) & (nl_S[:, 1] == 0) & (nl_S[:, 2] == 0))
-                    # When creating pairs between an atom and one of its periodic
-                    # images, the code generates multiple redundant pairs
-                    # (e.g. with shifts 0 1 1 and 0 -1 -1); and we want to only keep one
-                    # of these. We keep the pair in the positive half plane of shifts.
-                    | (
-                        (nl_S.sum(axis=1) < 0)
-                        | (
-                            (nl_S.sum(axis=1) == 0)
-                            & (
-                                (nl_S[:, 2] < 0)
-                                | ((nl_S[:, 2] == 0) & (nl_S[:, 1] < 0))
-                            )
-                        )
-                    )
-                )
-            )
-        )
-        selected = np.logical_not(reject_condition)
-        nl_i = nl_i[selected]
-        nl_j = nl_j[selected]
-        nl_S = nl_S[selected]
-        nl_D = nl_D[selected]
-
-    samples = np.concatenate([nl_i[:, None], nl_j[:, None], nl_S], axis=1)
-    distances = torch.from_numpy(nl_D).to(dtype=dtype, device=device)
-
-    return TensorBlock(
-        values=distances.reshape(-1, 3, 1),
-        samples=Labels(
-            names=[
-                "first_atom",
-                "second_atom",
-                "cell_shift_a",
-                "cell_shift_b",
-                "cell_shift_c",
-            ],
-            values=torch.from_numpy(samples).to(dtype=torch.int32, device=device),
-            assume_unique=True,
-        ),
-        components=[Labels.range("xyz", 3).to(device)],
-        properties=Labels.range("distance", 1).to(device),
-    )
-
-
 def _get_ase_input(
     atoms: ase.Atoms,
     name: str,
@@ -1091,7 +1007,7 @@ def _get_ase_input(
     tensor.set_info("quantity", infos["quantity"])
     tensor.set_info("unit", infos["unit"])
 
-    tensor.to(dtype=dtype, device=device)
+    tensor = tensor.to(dtype=dtype, device=device)
     return tensor
 
 
@@ -1099,7 +1015,7 @@ def _ase_to_torch_data(atoms, dtype, device):
     """Get the positions, cell and pbc from ASE atoms as torch tensors"""
 
     types = torch.from_numpy(atoms.numbers).to(dtype=torch.int32, device=device)
-    positions = torch.from_numpy(atoms.positions).to(dtype=dtype, device=device)
+    positions = torch.from_numpy(atoms.positions).to(dtype=dtype).to(device=device)
     cell = torch.zeros((3, 3), dtype=dtype, device=device)
     pbc = torch.tensor(atoms.pbc, dtype=torch.bool, device=device)
 
@@ -1191,10 +1107,13 @@ class SymmetrizedCalculator(ase.calculators.calculator.Calculator):
             self.quadrature_rotations, self.quadrature_weights = _get_quadrature(
                 lebedev_order, n_inplane_rotations, include_inversion
             )
-        else:
-            # no quadrature
-            self.quadrature_rotations = np.array([np.eye(3)])
-            self.quadrature_weights = np.array([1.0])
+        else:  # no quadrature
+            if include_inversion:  # identity and inversion
+                self.quadrature_rotations = np.array([np.eye(3), -np.eye(3)])
+                self.quadrature_weights = np.array([0.5, 0.5])
+            else:  # only the identity
+                self.quadrature_rotations = np.array([np.eye(3)])
+                self.quadrature_weights = np.array([1.0])
 
         self.batch_size = (
             batch_size if batch_size is not None else len(self.quadrature_rotations)
@@ -1272,40 +1191,13 @@ def _choose_quadrature(L_max: int) -> Tuple[int, int]:
     :param L_max: maximum spherical harmonic degree
     :return: (lebedev_order, n_inplane_rotations)
     """
+    # fmt: off
     available = [
-        3,
-        5,
-        7,
-        9,
-        11,
-        13,
-        15,
-        17,
-        19,
-        21,
-        23,
-        25,
-        27,
-        29,
-        31,
-        35,
-        41,
-        47,
-        53,
-        59,
-        65,
-        71,
-        77,
-        83,
-        89,
-        95,
-        101,
-        107,
-        113,
-        119,
-        125,
-        131,
+        3, 5, 7, 9, 11, 13, 15, 17, 19, 21, 23, 25, 27, 29, 31, 35, 41,
+        47, 53, 59, 65, 71, 77, 83, 89, 95, 101, 107, 113, 119, 125, 131,
     ]
+    # fmt: on
+
     # pick smallest order >= L_max
     n = min(o for o in available if o >= L_max)
     # minimal gamma count
@@ -1597,3 +1489,127 @@ def _average_over_group(
         out["stress"] = S_pg
 
     return out
+
+
+def _compute_requested_neighbors(
+    systems: List[System],
+    requested_options: List[NeighborListOptions],
+    check_consistency=False,
+) -> List[System]:
+    """
+    Compute all neighbor lists requested by ``model`` and store them inside the systems.
+    """
+    can_use_nvalchemi = HAS_NVALCHEMIOPS and all(
+        system.device.type == "cuda" for system in systems
+    )
+
+    if can_use_nvalchemi:
+        full_nl_options = []
+        half_nl_options = []
+        for options in requested_options:
+            if options.full_list:
+                full_nl_options.append(options)
+            else:
+                half_nl_options.append(options)
+
+        # Do the full neighbor lists with nvalchemi, and the rest with vesin
+        systems = _compute_requested_neighbors_nvalchemi(
+            systems=systems,
+            requested_options=full_nl_options,
+        )
+        systems = _compute_requested_neighbors_vesin(
+            systems=systems,
+            requested_options=half_nl_options,
+            check_consistency=check_consistency,
+        )
+    else:
+        systems = _compute_requested_neighbors_vesin(
+            systems=systems,
+            requested_options=requested_options,
+            check_consistency=check_consistency,
+        )
+
+    return systems
+
+
+def _compute_requested_neighbors_vesin(
+    systems: List[System],
+    requested_options: List[NeighborListOptions],
+    check_consistency=False,
+) -> List[System]:
+    """
+    Compute all neighbor lists requested by ``model`` and store them inside the systems,
+    using vesin.
+    """
+
+    system_devices = []
+    moved_systems = []
+    for system in systems:
+        system_devices.append(system.device)
+        if system.device.type not in ["cpu", "cuda"]:
+            moved_systems.append(system.to(device="cpu"))
+        else:
+            moved_systems.append(system)
+
+    vesin.metatomic.compute_requested_neighbors_from_options(
+        systems=moved_systems,
+        system_length_unit="angstrom",
+        options=requested_options,
+        check_consistency=check_consistency,
+    )
+
+    systems = []
+    for system, device in zip(moved_systems, system_devices, strict=True):
+        systems.append(system.to(device=device))
+
+    return systems
+
+
+def _compute_requested_neighbors_nvalchemi(systems, requested_options):
+    """
+    Compute all neighbor lists requested by ``model`` and store them inside the systems,
+    using nvalchemiops. This function should only be called if all systems are on CUDA
+    and all neighbor list options require a full neighbor list.
+    """
+
+    for options in requested_options:
+        assert options.full_list
+        for system in systems:
+            assert system.device.type == "cuda"
+
+            edge_index, _, S = nvalchemi_neighbor_list(
+                system.positions,
+                options.engine_cutoff("angstrom"),
+                cell=system.cell,
+                pbc=system.pbc,
+                return_neighbor_list=True,
+            )
+            D = (
+                system.positions[edge_index[1]]
+                - system.positions[edge_index[0]]
+                + S.to(system.cell.dtype) @ system.cell
+            )
+            P = edge_index.T
+
+            neighbors = TensorBlock(
+                D.reshape(-1, 3, 1),
+                samples=Labels(
+                    names=[
+                        "first_atom",
+                        "second_atom",
+                        "cell_shift_a",
+                        "cell_shift_b",
+                        "cell_shift_c",
+                    ],
+                    values=torch.hstack([P, S]),
+                ),
+                components=[
+                    Labels("xyz", torch.tensor([[0], [1], [2]], device=system.device))
+                ],
+                properties=Labels(
+                    "distance", torch.tensor([[0]], device=system.device)
+                ),
+            )
+            system.add_neighbor_list(options, neighbors)
+
+    return systems